verifies merged segment sizes in test
