+++ adapters/repos/db/lsmkv/compaction_set_integration_test.go
@@ -16,141 +16,24 @@ package lsmkv
 
 import (
 	"context"
-	"encoding/binary"
 	"fmt"
 	"math/rand"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"github.com/weaviate/sroar"
 	"github.com/weaviate/weaviate/entities/cyclemanager"
 )
 
-func compactionRoaringSetStrategy_Random(ctx context.Context, t *testing.T, opts []BucketOption) {
-	maxID := uint64(100)
-	maxElement := uint64(1e6)
-	iterations := uint64(100_000)
-
-	deleteRatio := 0.2   // 20% of all operations will be deletes, 80% additions
-	flushChance := 0.001 // on average one flush per 1000 iterations
-
-	r := getRandomSeed()
-
-	instr := generateRandomInstructions(r, maxID, maxElement, iterations, deleteRatio)
-	control := controlFromInstructions(instr, maxID)
-
-	b, err := NewBucket(ctx, t.TempDir(), "", nullLogger(), nil,
-		cyclemanager.NewCallbackGroupNoop(), cyclemanager.NewCallbackGroupNoop(), opts...)
-	require.Nil(t, err)
-
-	defer b.Shutdown(testCtx())
-
-	// so big it effectively never triggers as part of this test
-	b.SetMemtableThreshold(1e9)
-
-	compactions := 0
-	for _, inst := range instr {
-		key := make([]byte, 8)
-		binary.LittleEndian.PutUint64(key, inst.key)
-		if inst.addition {
-			b.RoaringSetAddOne(key, inst.element)
-		} else {
-			b.RoaringSetRemoveOne(key, inst.element)
-		}
-
-		if r.Float64() < flushChance {
-			require.Nil(t, b.FlushAndSwitch())
-
-			for b.disk.eligibleForCompaction() {
-				require.Nil(t, b.disk.compactOnce())
-				compactions++
-			}
-		}
-
-	}
-
-	// this is a sanity check to make sure the test setup actually does what we
-	// want. With the current setup, we expect on average to have ~100
-	// compactions. It would be extremely unexpected to have fewer than 25.
-	assert.Greater(t, compactions, 25)
-
-	verifyBucketAgainstControl(t, b, control)
-}
-
-func verifyBucketAgainstControl(t *testing.T, b *Bucket, control []*sroar.Bitmap) {
-	// This test was built before the bucket had cursors, so we are retrieving
-	// each key individually, rather than cursing over the entire bucket.
-	// However, this is also good for isolation purposes, this test tests
-	// compactions, not cursors.
-
-	for i, controlBM := range control {
-		key := make([]byte, 8)
-		binary.LittleEndian.PutUint64(key, uint64(i))
-
-		actual, err := b.RoaringSetGet(key)
-		require.Nil(t, err)
-
-		assert.Equal(t, controlBM.ToArray(), actual.ToArray())
-
-	}
-}
-
-type roaringSetInstruction struct {
-	// is a []byte in reality, but makes the test setup easier if we pretent
-	// its an int
-	key     uint64
-	element uint64
-
-	// true=addition, false=deletion
-	addition bool
-}
-
-func generateRandomInstructions(r *rand.Rand, maxID, maxElement, iterations uint64,
-	deleteRatio float64,
-) []roaringSetInstruction {
-	instr := make([]roaringSetInstruction, iterations)
-
-	for i := range instr {
-		instr[i].key = uint64(r.Intn(int(maxID)))
-		instr[i].element = uint64(r.Intn(int(maxElement)))
-
-		if r.Float64() > deleteRatio {
-			instr[i].addition = true
-		} else {
-			instr[i].addition = false
-		}
-	}
-
-	return instr
-}
-
-func controlFromInstructions(instr []roaringSetInstruction, maxID uint64) []*sroar.Bitmap {
-	out := make([]*sroar.Bitmap, maxID)
-	for i := range out {
-		out[i] = sroar.NewBitmap()
-	}
-
-	for _, inst := range instr {
-		if inst.addition {
-			out[inst.key].Set(inst.element)
-		} else {
-			out[inst.key].Remove(inst.element)
-		}
-	}
-
-	return out
-}
-
-func compactionRoaringSetStrategy(ctx context.Context, t *testing.T, opts []BucketOption,
+func compactionSetStrategy(ctx context.Context, t *testing.T, opts []BucketOption,
 	expectedMinSize, expectedMaxSize int64,
 ) {
 	size := 100
 
 	type kv struct {
-		key       []byte
-		additions []uint64
-		deletions []uint64
+		key    []byte
+		values [][]byte
+		delete bool
 	}
 	// this segment is not part of the merge, but might still play a role in
 	// overall results. For example if one of the later segments has a tombstone
@@ -178,78 +61,81 @@ func compactionRoaringSetStrategy(ctx context.Context, t *testing.T, opts []Buck
 		// 7.) present in an unrelated previous segment, never touched again
 		for i := 0; i < size; i++ {
 			key := []byte(fmt.Sprintf("key-%02d", i))
-			value1 := uint64(i) + 1
-			value2 := uint64(i) + 2
-			values := []uint64{value1, value2}
+
+			value1 := []byte(fmt.Sprintf("value-%02d-01", i))
+			value2 := []byte(fmt.Sprintf("value-%02d-02", i))
+			values := [][]byte{value1, value2}
 
 			switch i % 8 {
 			case 0:
 				// add to segment 1
 				segment1 = append(segment1, kv{
-					key:       key,
-					additions: values[:1],
+					key:    key,
+					values: values[:1],
 				})
 
 				// leave this element untouched in the second segment
 				expected = append(expected, kv{
-					key:       key,
-					additions: values[:1],
+					key:    key,
+					values: values[:1],
 				})
 
 			case 1:
 				// add to segment 1
 				segment1 = append(segment1, kv{
-					key:       key,
-					additions: values[:1],
+					key:    key,
+					values: values[:1],
 				})
 
 				// update in the second segment
 				segment2 = append(segment2, kv{
-					key:       key,
-					additions: values[1:],
+					key:    key,
+					values: values[1:2],
 				})
 
 				expected = append(expected, kv{
-					key:       key,
-					additions: values,
+					key:    key,
+					values: values,
 				})
 
 			case 2:
 				// add both to segment 1, delete the first
 				segment1 = append(segment1, kv{
-					key:       key,
-					additions: values,
+					key:    key,
+					values: values,
 				})
 
 				// delete first element in the second segment
 				segment2 = append(segment2, kv{
-					key:       key,
-					deletions: values[:1],
+					key:    key,
+					values: values[:1],
+					delete: true,
 				})
 
 				// only the 2nd element should be left in the expected
 				expected = append(expected, kv{
-					key:       key,
-					additions: values[1:],
+					key:    key,
+					values: values[1:2],
 				})
 
 			case 3:
 				// add both to segment 1, delete the second
 				segment1 = append(segment1, kv{
-					key:       key,
-					additions: values,
+					key:    key,
+					values: values,
 				})
 
 				// delete second element in the second segment
 				segment2 = append(segment2, kv{
-					key:       key,
-					deletions: values[1:],
+					key:    key,
+					values: values[1:],
+					delete: true,
 				})
 
 				// only the 1st element should be left in the expected
 				expected = append(expected, kv{
-					key:       key,
-					additions: values[:1],
+					key:    key,
+					values: values[:1],
 				})
 
 			case 4:
@@ -257,30 +143,36 @@ func compactionRoaringSetStrategy(ctx context.Context, t *testing.T, opts []Buck
 
 				// only add to segment 2 (first entry)
 				segment2 = append(segment2, kv{
-					key:       key,
-					additions: values,
+					key:    key,
+					values: values,
 				})
 
 				expected = append(expected, kv{
-					key:       key,
-					additions: values,
+					key:    key,
+					values: values,
 				})
 
 			case 5:
 				// only part of a previous segment, which is not part of the merge
 				previous1 = append(previous1, kv{
-					key:       key,
-					additions: values[:1],
+					key:    key,
+					values: values[:1],
 				})
 				previous2 = append(previous2, kv{
-					key:       key,
-					additions: values[1:],
+					key:    key,
+					values: values[1:],
 				})
 
 				// delete in segment 1
 				segment1 = append(segment1, kv{
-					key:       key,
-					deletions: values,
+					key:    key,
+					values: values[:1],
+					delete: true,
+				})
+				segment1 = append(segment1, kv{
+					key:    key,
+					values: values[1:],
+					delete: true,
 				})
 
 				// should not have any values in expected at all, not even a key
@@ -288,18 +180,24 @@ func compactionRoaringSetStrategy(ctx context.Context, t *testing.T, opts []Buck
 			case 6:
 				// only part of a previous segment, which is not part of the merge
 				previous1 = append(previous1, kv{
-					key:       key,
-					additions: values[:1],
+					key:    key,
+					values: values[:1],
 				})
 				previous2 = append(previous2, kv{
-					key:       key,
-					additions: values[1:],
+					key:    key,
+					values: values[1:],
 				})
 
 				// delete in segment 2
 				segment2 = append(segment2, kv{
-					key:       key,
-					deletions: values,
+					key:    key,
+					values: values[:1],
+					delete: true,
+				})
+				segment2 = append(segment2, kv{
+					key:    key,
+					values: values[1:],
+					delete: true,
 				})
 
 				// should not have any values in expected at all, not even a key
@@ -307,17 +205,17 @@ func compactionRoaringSetStrategy(ctx context.Context, t *testing.T, opts []Buck
 			case 7:
 				// part of a previous segment
 				previous1 = append(previous1, kv{
-					key:       key,
-					additions: values[:1],
+					key:    key,
+					values: values[:1],
 				})
 				previous2 = append(previous2, kv{
-					key:       key,
-					additions: values[1:],
+					key:    key,
+					values: values[1:],
 				})
 
 				expected = append(expected, kv{
-					key:       key,
-					additions: values,
+					key:    key,
+					values: values,
 				})
 			}
 		}
@@ -345,65 +243,63 @@ func compactionRoaringSetStrategy(ctx context.Context, t *testing.T, opts []Buck
 	})
 
 	t.Run("import and flush previous segments", func(t *testing.T) {
-		for _, kv := range previous1 {
-			err := bucket.RoaringSetAddList(kv.key, kv.additions)
-			require.NoError(t, err)
+		for _, pair := range previous1 {
+			err := bucket.SetAdd(pair.key, pair.values)
+			require.Nil(t, err)
 		}
 
-		require.NoError(t, bucket.FlushAndSwitch())
+		require.Nil(t, bucket.FlushAndSwitch())
 
-		for _, kv := range previous2 {
-			err := bucket.RoaringSetAddList(kv.key, kv.additions)
-			require.NoError(t, err)
+		for _, pair := range previous2 {
+			err := bucket.SetAdd(pair.key, pair.values)
+			require.Nil(t, err)
 		}
 
-		require.NoError(t, bucket.FlushAndSwitch())
+		require.Nil(t, bucket.FlushAndSwitch())
 	})
 
 	t.Run("import segment 1", func(t *testing.T) {
-		for _, kv := range segment1 {
-			if len(kv.additions) > 0 {
-				err := bucket.RoaringSetAddList(kv.key, kv.additions)
-				require.NoError(t, err)
-			}
-			for i := range kv.deletions {
-				err := bucket.RoaringSetRemoveOne(kv.key, kv.deletions[i])
-				require.NoError(t, err)
+		for _, pair := range segment1 {
+			if !pair.delete {
+				err := bucket.SetAdd(pair.key, pair.values)
+				require.Nil(t, err)
+			} else {
+				err := bucket.SetDeleteSingle(pair.key, pair.values[0])
+				require.Nil(t, err)
 			}
 		}
 	})
 
 	t.Run("flush to disk", func(t *testing.T) {
-		require.NoError(t, bucket.FlushAndSwitch())
+		require.Nil(t, bucket.FlushAndSwitch())
 	})
 
 	t.Run("import segment 2", func(t *testing.T) {
-		for _, kv := range segment2 {
-			if len(kv.additions) > 0 {
-				err := bucket.RoaringSetAddList(kv.key, kv.additions)
-				require.NoError(t, err)
-			}
-			for i := range kv.deletions {
-				err := bucket.RoaringSetRemoveOne(kv.key, kv.deletions[i])
-				require.NoError(t, err)
+		for _, pair := range segment2 {
+			if !pair.delete {
+				err := bucket.SetAdd(pair.key, pair.values)
+				require.Nil(t, err)
+			} else {
+				err := bucket.SetDeleteSingle(pair.key, pair.values[0])
+				require.Nil(t, err)
 			}
 		}
 	})
 
 	t.Run("flush to disk", func(t *testing.T) {
-		require.NoError(t, bucket.FlushAndSwitch())
+		require.Nil(t, bucket.FlushAndSwitch())
 	})
 
 	t.Run("verify control before compaction", func(t *testing.T) {
 		var retrieved []kv
 
-		c := bucket.CursorRoaringSet()
+		c := bucket.SetCursor()
 		defer c.Close()
 
 		for k, v := c.First(); k != nil; k, v = c.Next() {
 			retrieved = append(retrieved, kv{
-				key:       k,
-				additions: v.ToArray(),
+				key:    k,
+				values: v,
 			})
 		}
 
@@ -423,7 +319,7 @@ func compactionRoaringSetStrategy(ctx context.Context, t *testing.T, opts []Buck
 				// segment1 and segment2 merged
 				// none of them is root segment, so tombstones
 				// will not be removed regardless of keepTombstones setting
-				assertSecondSegmentOfSize(t, bucket, 26768, 26768)
+				assertSecondSegmentOfSize(t, bucket, 8556, 8556)
 			}
 		}
 	})
@@ -431,13 +327,13 @@ func compactionRoaringSetStrategy(ctx context.Context, t *testing.T, opts []Buck
 	t.Run("verify control after compaction", func(t *testing.T) {
 		var retrieved []kv
 
-		c := bucket.CursorRoaringSet()
+		c := bucket.SetCursor()
 		defer c.Close()
 
 		for k, v := c.First(); k != nil; k, v = c.Next() {
 			retrieved = append(retrieved, kv{
-				key:       k,
-				additions: v.ToArray(),
+				key:    k,
+				values: v,
 			})
 		}
 
@@ -446,7 +342,7 @@ func compactionRoaringSetStrategy(ctx context.Context, t *testing.T, opts []Buck
 	})
 }
 
-func compactionRoaringSetStrategy_RemoveUnnecessary(ctx context.Context, t *testing.T, opts []BucketOption) {
+func compactionSetStrategy_RemoveUnnecessary(ctx context.Context, t *testing.T, opts []BucketOption) {
 	// in this test each segment reverses the action of the previous segment so
 	// that in the end a lot of information is present in the individual segments
 	// which is no longer needed. We then verify that after all compaction this
@@ -455,7 +351,7 @@ func compactionRoaringSetStrategy_RemoveUnnecessary(ctx context.Context, t *test
 
 	type kv struct {
 		key    []byte
-		values []uint64
+		values [][]byte
 	}
 
 	key := []byte("my-key")
@@ -464,7 +360,7 @@ func compactionRoaringSetStrategy_RemoveUnnecessary(ctx context.Context, t *test
 	dirName := t.TempDir()
 
 	t.Run("init bucket", func(t *testing.T) {
-		b, err := NewBucket(ctx, dirName, "", nullLogger(), nil,
+		b, err := NewBucket(ctx, dirName, dirName, nullLogger(), nil,
 			cyclemanager.NewCallbackGroupNoop(), cyclemanager.NewCallbackGroupNoop(), opts...)
 		require.Nil(t, err)
 
@@ -478,14 +374,16 @@ func compactionRoaringSetStrategy_RemoveUnnecessary(ctx context.Context, t *test
 		for i := 0; i < size; i++ {
 			if i != 0 {
 				// we can only delete an existing value if this isn't the first write
-				err := bucket.RoaringSetRemoveOne(key, uint64(i)-1)
-				require.NoError(t, err)
+				value := []byte(fmt.Sprintf("value-%05d", i-1))
+				err := bucket.SetDeleteSingle(key, value)
+				require.Nil(t, err)
 			}
 
-			err := bucket.RoaringSetAddOne(key, uint64(i))
-			require.NoError(t, err)
+			value := []byte(fmt.Sprintf("value-%05d", i))
+			err := bucket.SetAdd(key, [][]byte{value})
+			require.Nil(t, err)
 
-			require.NoError(t, bucket.FlushAndSwitch())
+			require.Nil(t, bucket.FlushAndSwitch())
 		}
 	})
 
@@ -494,17 +392,17 @@ func compactionRoaringSetStrategy_RemoveUnnecessary(ctx context.Context, t *test
 		expected := []kv{
 			{
 				key:    key,
-				values: []uint64{uint64(size) - 1},
+				values: [][]byte{[]byte(fmt.Sprintf("value-%05d", size-1))},
 			},
 		}
 
-		c := bucket.CursorRoaringSet()
+		c := bucket.SetCursor()
 		defer c.Close()
 
 		for k, v := c.First(); k != nil; k, v = c.Next() {
 			retrieved = append(retrieved, kv{
 				key:    k,
-				values: v.ToArray(),
+				values: v,
 			})
 		}
 
@@ -526,17 +424,17 @@ func compactionRoaringSetStrategy_RemoveUnnecessary(ctx context.Context, t *test
 		expected := []kv{
 			{
 				key:    key,
-				values: []uint64{uint64(size) - 1},
+				values: [][]byte{[]byte(fmt.Sprintf("value-%05d", size-1))},
 			},
 		}
 
-		c := bucket.CursorRoaringSet()
+		c := bucket.SetCursor()
 		defer c.Close()
 
 		for k, v := c.First(); k != nil; k, v = c.Next() {
 			retrieved = append(retrieved, kv{
 				key:    k,
-				values: v.ToArray(),
+				values: v,
 			})
 		}
 
@@ -544,7 +442,7 @@ func compactionRoaringSetStrategy_RemoveUnnecessary(ctx context.Context, t *test
 	})
 }
 
-func compactionRoaringSetStrategy_FrequentPutDeleteOperations(ctx context.Context, t *testing.T, opts []BucketOption) {
+func compactionSetStrategy_FrequentPutDeleteOperations(ctx context.Context, t *testing.T, opts []BucketOption) {
 	// In this test we are testing that the compaction works well for set collection
 	maxSize := 10
 
@@ -553,14 +451,14 @@ func compactionRoaringSetStrategy_FrequentPutDeleteOperations(ctx context.Contex
 			var bucket *Bucket
 
 			key := []byte("key-original")
-			value1 := uint64(1)
-			value2 := uint64(2)
-			values := []uint64{value1, value2}
+			value1 := []byte("value-01")
+			value2 := []byte("value-02")
+			values := [][]byte{value1, value2}
 
 			dirName := t.TempDir()
 
 			t.Run("init bucket", func(t *testing.T) {
-				b, err := NewBucket(ctx, dirName, "", nullLogger(), nil,
+				b, err := NewBucket(ctx, dirName, dirName, nullLogger(), nil,
 					cyclemanager.NewCallbackGroupNoop(), cyclemanager.NewCallbackGroupNoop(), opts...)
 				require.Nil(t, err)
 
@@ -572,24 +470,24 @@ func compactionRoaringSetStrategy_FrequentPutDeleteOperations(ctx context.Contex
 
 			t.Run("import and flush segments", func(t *testing.T) {
 				for i := 0; i < size; i++ {
-					err := bucket.RoaringSetAddList(key, values)
+					err := bucket.SetAdd(key, values)
 					require.Nil(t, err)
 
 					if size == 5 {
 						// delete all
-						err := bucket.RoaringSetRemoveOne(key, values[0])
+						err := bucket.SetDeleteSingle(key, values[0])
 						require.Nil(t, err)
-						err = bucket.RoaringSetRemoveOne(key, values[1])
+						err = bucket.SetDeleteSingle(key, values[1])
 						require.Nil(t, err)
 					} else if size == 6 {
 						// delete only one value
-						err := bucket.RoaringSetRemoveOne(key, values[0])
+						err := bucket.SetDeleteSingle(key, values[0])
 						require.Nil(t, err)
 					} else if i != size-1 {
 						// don't delete from the last segment
-						err := bucket.RoaringSetRemoveOne(key, values[0])
+						err := bucket.SetDeleteSingle(key, values[0])
 						require.Nil(t, err)
-						err = bucket.RoaringSetRemoveOne(key, values[1])
+						err = bucket.SetDeleteSingle(key, values[1])
 						require.Nil(t, err)
 					}
 
@@ -598,14 +496,14 @@ func compactionRoaringSetStrategy_FrequentPutDeleteOperations(ctx context.Contex
 			})
 
 			t.Run("verify that objects exist before compaction", func(t *testing.T) {
-				res, err := bucket.RoaringSetGet(key)
-				require.NoError(t, err)
+				res, err := bucket.SetList(key)
+				assert.Nil(t, err)
 				if size == 5 {
-					assert.Equal(t, 0, res.GetCardinality())
+					assert.Len(t, res, 0)
 				} else if size == 6 {
-					assert.Equal(t, 1, res.GetCardinality())
+					assert.Len(t, res, 1)
 				} else {
-					assert.Equal(t, 2, res.GetCardinality())
+					assert.Len(t, res, 2)
 				}
 			})
 
@@ -615,19 +513,19 @@ func compactionRoaringSetStrategy_FrequentPutDeleteOperations(ctx context.Contex
 
 			t.Run("compact until no longer eligible", func(t *testing.T) {
 				for bucket.disk.eligibleForCompaction() {
-					require.NoError(t, bucket.disk.compactOnce())
+					require.Nil(t, bucket.disk.compactOnce())
 				}
 			})
 
 			t.Run("verify that objects exist after compaction", func(t *testing.T) {
-				res, err := bucket.RoaringSetGet(key)
-				require.NoError(t, err)
+				res, err := bucket.SetList(key)
+				assert.Nil(t, err)
 				if size == 5 {
-					assert.Equal(t, 0, res.GetCardinality())
+					assert.Len(t, res, 0)
 				} else if size == 6 {
-					assert.Equal(t, 1, res.GetCardinality())
+					assert.Len(t, res, 1)
 				} else {
-					assert.Equal(t, 2, res.GetCardinality())
+					assert.Len(t, res, 2)
 				}
 			})
 		})
